22.09.2019 20:49

+++ src/main/java/packModel/Product.java
@@ -17,7 +17,7 @@ public class Product implements IbaseEntity {
     @GeneratedValue(strategy = GenerationType.AUTO)
     private Long id;
     private String name;
-    private int price;
+    private double price;
 
     @Formula(value = "(case when (taxType='PRODUCT') then (price*0.23) when (taxType='SERVICE') then (price*0.08) end)")
     private double tax;
@@ -31,7 +31,7 @@ public class Product implements IbaseEntity {
     @ManyToOne()
     private Invoice invoice;
 
-    public Product(String name, int price, TaxType taxType, int stock) {
+    public Product(String name, double price, TaxType taxType, int stock) {
         this.name = name;
         this.price = price;
         this.taxType = taxType;
